Move httpsRequest into utils file
